requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==11.3.0
pyautogui==0.9.54
